run non hub tests with updated config
