+++ .azure-devops/templates/nightly-tests.yml
@@ -37,28 +37,28 @@ steps:
     env:
       COVERAGE_FILE: .coverage.all
 
-  - task: AzureCLI@2
-    continueOnError: true
-    displayName: 'IoTHub integration tests - devices, messaging and modules'
-    inputs:
-      azureSubscription: az_cli_nightly_conn
-      scriptType: bash
-      scriptLocation: inlineScript
-      inlineScript: |
-        export COVERAGE_FILE=.coverage.all
-        pytest -vv azext_iot/tests/iothub/devices azext_iot/tests/iothub/messaging azext_iot/tests/iothub/modules -k "_int.py" --dist=loadfile -n ${{ parameters.num_threads }} --reruns ${{ parameters.num_reruns }} --reruns-delay ${{ parameters.reruns_delay }} --cov=azext_iot --cov-config .coveragerc --junitxml=junit/test-iotext-int.xml
+  # - task: AzureCLI@2
+  #   continueOnError: true
+  #   displayName: 'IoTHub integration tests - devices, messaging and modules'
+  #   inputs:
+  #     azureSubscription: az_cli_nightly_conn
+  #     scriptType: bash
+  #     scriptLocation: inlineScript
+  #     inlineScript: |
+  #       export COVERAGE_FILE=.coverage.all
+  #       pytest -vv azext_iot/tests/iothub/devices azext_iot/tests/iothub/messaging azext_iot/tests/iothub/modules -k "_int.py" --dist=loadfile -n ${{ parameters.num_threads }} --reruns ${{ parameters.num_reruns }} --reruns-delay ${{ parameters.reruns_delay }} --cov=azext_iot --cov-config .coveragerc --junitxml=junit/test-iotext-int.xml
 
 
-  - task: AzureCLI@2
-    continueOnError: true
-    displayName: 'IoTHub integration tests - jobs, hub and configuration'
-    inputs:
-      azureSubscription: az_cli_nightly_conn
-      scriptType: bash
-      scriptLocation: inlineScript
-      inlineScript: |
-        export COVERAGE_FILE=.coverage.all
-        pytest -vv azext_iot/tests/iothub --ignore=azext_iot/tests/iothub/devices --ignore=azext_iot/tests/iothub/messaging --ignore=azext_iot/tests/iothub/modules -k "_int.py" --dist=loadfile -n ${{ parameters.num_threads }} --reruns ${{ parameters.num_reruns }} --reruns-delay ${{ parameters.reruns_delay }} --cov=azext_iot --cov-config .coveragerc --junitxml=junit/test-iotext-int.xml --deselect=azext_iot/tests/iothub/test_iothub_discovery_int.py::TestIoTHubDiscovery::test_iothub_targets
+  # - task: AzureCLI@2
+  #   continueOnError: true
+  #   displayName: 'IoTHub integration tests - jobs, hub and configuration'
+  #   inputs:
+  #     azureSubscription: az_cli_nightly_conn
+  #     scriptType: bash
+  #     scriptLocation: inlineScript
+  #     inlineScript: |
+  #       export COVERAGE_FILE=.coverage.all
+  #       pytest -vv azext_iot/tests/iothub --ignore=azext_iot/tests/iothub/devices --ignore=azext_iot/tests/iothub/messaging --ignore=azext_iot/tests/iothub/modules -k "_int.py" --dist=loadfile -n ${{ parameters.num_threads }} --reruns ${{ parameters.num_reruns }} --reruns-delay ${{ parameters.reruns_delay }} --cov=azext_iot --cov-config .coveragerc --junitxml=junit/test-iotext-int.xml --deselect=azext_iot/tests/iothub/test_iothub_discovery_int.py::TestIoTHubDiscovery::test_iothub_targets
 
   - task: AzureCLI@2
     continueOnError: true
